1_Supermarket_Sales_Analysis/deployment/requirements.txt
streamlit
pandas
numpy
scipy
plotly.express
datetime
seaborn
matplotlib
